refactor(idea_service): extract resonance + concept token helpers (#163)
  before: 2423 lines
  after:  2347 lines (-76)

321 tests pass.
